Removed old implementation and minified final commonjs bundle with uglify
