+++ package.json
@@ -61,6 +61,7 @@
 		"rollup-plugin-replace": "^2.2.0",
 		"rollup-plugin-sourcemaps": "^0.4.2",
 		"rollup-plugin-typescript2": "^0.22.1",
+		"rollup-plugin-uglify": "^6.0.4",
 		"rollup-watch": "^4.3.1",
 		"sass-loader": "^8.0.2",
 		"ts-loader": "^6.2.1",
