+++ src/components/Centered.tsx
@@ -1,6 +1,3 @@
 import * as React from "react";
-import "../styles.css";
 
 export const Centered: React.FC = (props) => <div className={`spaces-centered`}>{props.children}</div>;
-
-export const CenteredVertically: React.FC = (props) => <div className={`spaces-centered-vertically`}>{props.children}</div>;
