--- src/components/experimental/Centered.tsx
@@ -1,3 +1,0 @@
-import * as React from "react";
-
-export const Centered: React.FC = (props) => <div className={`spaces-centered`}>{props.children}</div>;
