--- src/components/experimental/Fixed.tsx
@@ -1,14 +1,0 @@
-import { ICommonProps, SizeUnit, Type } from "../../core-types";
-import * as React from "react";
-import { Space } from "./Space";
-
-interface IFixedProps extends ICommonProps {
-	width?: SizeUnit;
-	height: SizeUnit;
-}
-
-export const Fixed: React.FC<IFixedProps> = ({ width, height, children, ...commonProps }) => (
-	<Space {...commonProps} type={Type.Fixed} position={{ width: width, height: height }}>
-		{children}
-	</Space>
-);
