--- src/components/experimental/Layer.tsx
@@ -1,4 +1,0 @@
-import * as React from "react";
-import { LayerContext } from "../../core-react";
-
-export const Layer: React.FC<{ zIndex: number }> = (props) => <LayerContext.Provider value={props.zIndex}>{props.children}</LayerContext.Provider>;
