+++ src/components/index.tsx
@@ -1,4 +1,4 @@
-import "../../styles.css";
+import "../styles.css";
 export * from "./Anchored";
 export * from "./Centered";
 export * from "./CenteredVertically";
@@ -9,4 +9,4 @@ export * from "./Layer";
 export * from "./Positioned";
 export * from "./SpaceInfo";
 export * from "./ViewPort";
-export { CenterType } from "../../core-types";
+export { CenterType } from "../core-types";
