+++ src/components/stories/00-docs/01-GettingStarted.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { CommonHeader, DemoUI } from "../Utils";
-import { ViewPort, Fixed, Top, Right, LeftResizable, BottomResizable, Fill } from "../../experimental";
+import { ViewPort, Fixed, Top, Right, LeftResizable, BottomResizable, Fill } from "../../";
 import { CenterType } from "../../../types";
 
 <CommonHeader />
