+++ src/components/stories/02-components/CenteredVertically.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, CenteredVertically } from "../../experimental";
+import { ViewPort, CenteredVertically } from "../../";
 import { blue } from "../Utils";
 import { CommonHeader } from "../Utils";
