+++ src/components/stories/02-components/Custom.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, Custom, AnchorType } from "../../experimental";
+import { ViewPort, Custom, AnchorType } from "../../";
 import { green, description } from "../Utils";
 import { CommonHeader, PropsTable, PropsHeader, Prop, StandardProps } from "../Utils";
