+++ src/components/stories/02-components/Layer.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, Experimental, Layer, Centered, Positioned } from "../../experimental";
+import { ViewPort, Experimental, Layer, Centered, Positioned } from "../../";
 import { blue, red } from "../Utils";
 import { CommonHeader, description, PropsTable, Prop } from "../Utils";
