+++ src/components/stories/02-components/Positioned.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, Centered, Positioned } from "../../experimental";
-import { useParentSpace } from "../../../hooks/useParentSpace";
+import { ViewPort, Centered, Positioned } from "../../";
+//import { useParentSpace } from "../../../hooks/useParentSpace";
 import { green, description } from "../Utils";
 import { CommonHeader, PropsTable, PropsHeader, Prop, StandardProps } from "../Utils";
 
@@ -43,12 +43,12 @@ Absolutely positioned space within it's parent.
 	<Story name="With drag handle">
 		{() => {
 			const DragHandle = () => {
-				const parent = useParentSpace();
+				//const parent = useParentSpace();
 				return (
 					<Centered>
-						<button onMouseDown={(e) => parent.startMouseDrag(e)}>
+						{/* <button onMouseDown={(e) => parent.startMouseDrag(e)}>
 							Drag handle
-						</button>
+						</button> */}
 					</Centered>
 				)
 			}
