+++ src/components/stories/02-components/TopResizable.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, TopResizable } from "../../experimental";
+import { ViewPort, TopResizable } from "../../";
 import { green, description, lorem } from "../Utils";
 import { CommonHeader, PropsTable, StandardProps, AnchoredProps, ResizableProps } from "../Utils";
