+++ src/components/stories/02-components/ViewPort.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Preview, Props } from "@storybook/addon-docs/blocks";
 import { withKnobs } from "@storybook/addon-knobs";
-import { ViewPort, LeftResizable, TopResizable, RightResizable, BottomResizable, Fill } from "../../experimental";
+import { ViewPort, LeftResizable, TopResizable, RightResizable, BottomResizable, Fill } from "../../";
 import { red, green, blue, description } from "../Utils";
 import { CommonHeader, PropsTable, PropsHeader, Prop } from "../Utils";
