--- src/experimental.ts
@@ -1,24 +1,0 @@
-export {
-	AnchorType,
-	CenterType,
-	Fill,
-	Left,
-	Right,
-	Top,
-	Bottom,
-	LeftResizable,
-	RightResizable,
-	TopResizable,
-	BottomResizable,
-	Positioned,
-	Custom,
-	Fixed,
-	ViewPort,
-	Layer,
-	Centered,
-	CenteredVertically,
-} from "./components/experimental";
-
-//export { useParentSpace } from "./hooks/useParentSpace";
-
-//export { ResizeType, ISpaceInfo, IParentSpace, IPosition } from "./types";
