+++ src/index.ts
@@ -1,11 +1,24 @@
-export * from "./components/Centered";
-export * from "./components/ClearFix";
-export * from "./components/Fixed";
-export * from "./components/ViewPort";
-export * from "./components/Layer";
-export * from "./components/SpaceInfo";
-export { Fill, Left, Right, Top, Bottom, LeftResizable, RightResizable, TopResizable, BottomResizable, Positioned, Custom } from "./components/Space";
+export {
+	AnchorType,
+	CenterType,
+	Fill,
+	Left,
+	Right,
+	Top,
+	Bottom,
+	LeftResizable,
+	RightResizable,
+	TopResizable,
+	BottomResizable,
+	Positioned,
+	Custom,
+	Fixed,
+	ViewPort,
+	Layer,
+	Centered,
+	CenteredVertically,
+} from "./components";
 
-export { useParentSpace } from "./hooks/useParentSpace";
+//export { useParentSpace } from "./hooks/useParentSpace";
 
-export { CenterType, AnchorType, ResizeType, ISpaceInfo, IParentSpace, IPosition } from "./types";
+//export { ResizeType, ISpaceInfo, IParentSpace, IPosition } from "./types";
